Closed #26 - now supporting +{} expressions. While {A} means zero or more A's, +{A} means one or more A's.
The implementation is tricky, because {A} is in practice a Closure that returns 'A' when called.

When parsing B >> +{A}, the + unary operator operates on a Closure. Precedent-wise, the unary plus happens before the closure is called, so the plus operator should accept a closure and return a closure. But the returned closure should know it reflects a one-or-more rather than a zero-or-more...

To achieve that, I am using a class (AtLeastOneClosure) which extends Closure. So +{A} is an AtLeastOneClosure that references {A}. Calling +{A} returns A. But, when instantiating a Repeated object, it is possible to check if the closure is an AtLeastOneClosure, or just a regular closure. This is how Repeated is set to be one-or-more or zero-or-more.