Closed #27 - an ExPath is now an ExPath<CustomInfo> - it may hold additional information about the path, for example - the exact math point, if the path leads to a subexpression and the match itself is somewhere inside the subexpression.
match() has changed to get a closure that returns the info, and not merely TRUE when it finds a match. This way, the info is propagated up and stored in the ExPath.

match() does not support multiple info objects returned from a match point. It could be handy - if you may find multiple matches in the same subexpression the path leads to. This support may be added in the future.