Closed #28 - demonstrating the transformation of
A & {"," & A}

into

+{A}/","

First we match the pattern A & {"," & A} wherever it appears (even when it appears multiple times in an expression, even if one occurrence wraps the other...).

Only the first match is transformed each time, because the transformation may invalidate other matches.

Transformation means making the changes to the matched expression, but also to its children. Then, if the transformation made a Multinary expression have only one child, the single children is moved to be a child of the parent of the transformed expression...